LayoutTests/svg/css/parse-length-expected.txt

CONSOLE MESSAGE: Error: Invalid value for <rect> attribute width="auto"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute width="  100"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute width="100   "
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute x="  100"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute x="100   "
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute y="  100"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute y="100   "
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute width="100   px"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute width="100px;"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute width="100px !important"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute width="{ 100px }"
CONSOLE MESSAGE: Error: Invalid negative value for <rect> attribute width="-100px"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute x="auto"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute x="100   px"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute x="100px;"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute x="100px !important"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute x="{ 100px }"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute y="auto"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute y="100   px"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute y="100px;"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute y="100px !important"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute y="{ 100px }"
Test that 'with' presentation attribute is parsed with CSS presentation rules.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS computedStyle("width", "auto") is "auto"
PASS computedStyle("width", "  100") is "100px"
PASS computedStyle("width", "100   ") is "100px"
PASS computedStyle("width", "100px") is "100px"
PASS computedStyle("width", "1em") is "16px"
PASS computedStyle("width", "1ex") is "12.800000190734863px"
PASS computedStyle("width", "20%") is "20%"
PASS computedStyle("x", "  100") is "100px"
PASS computedStyle("x", "100   ") is "100px"
PASS computedStyle("x", "100px") is "100px"
PASS computedStyle("x", "1em") is "16px"
PASS computedStyle("x", "1ex") is "12.800000190734863px"
PASS computedStyle("x", "20%") is "20%"
PASS computedStyle("x", "-200px") is "-200px"
PASS computedStyle("y", "  100") is "100px"
PASS computedStyle("y", "100   ") is "100px"
PASS computedStyle("y", "100px") is "100px"
PASS computedStyle("y", "1em") is "16px"
PASS computedStyle("y", "1ex") is "12.800000190734863px"
PASS computedStyle("y", "20%") is "20%"
PASS computedStyle("y", "-200px") is "-200px"
PASS computedStyle("width", "100   px") is "auto"
PASS computedStyle("width", "100px;") is "auto"
PASS computedStyle("width", "100px !important") is "auto"
PASS computedStyle("width", "{ 100px }") is "auto"
PASS computedStyle("width", "-100px") is "auto"
PASS computedStyle("x", "auto") is "0px"
PASS computedStyle("x", "100   px") is "0px"
PASS computedStyle("x", "100px;") is "0px"
PASS computedStyle("x", "100px !important") is "0px"
PASS computedStyle("x", "{ 100px }") is "0px"
PASS computedStyle("y", "auto") is "0px"
PASS computedStyle("y", "100   px") is "0px"
PASS computedStyle("y", "100px;") is "0px"
PASS computedStyle("y", "100px !important") is "0px"
PASS computedStyle("y", "{ 100px }") is "0px"
PASS successfullyParsed is true

TEST COMPLETE
